Added chapter 2 complete code
