+++ go/chapter02/question2.2/question2_2.go
@@ -15,45 +15,48 @@ Algorithm KTH_FROM_LAST(list,k):
 package main
 
 import (
-	"container/list"
 	"fmt"
+	"go/chapter02-linkedlists/list"
 )
-var sMap map[int]bool
+
 func main() {
 	l := list.New()
 	for i := 1; i < 100; i++ {
 		l.PushBack(i)
 	}
- 	kFromLastElem := findKFromLast(l, 3)
-    fmt.Println("Iterative Method : ",kFromLastElem.Value)
+	kFromLastElem := findKFromLast(l, 3)
+	fmt.Println("Iterative Method : ", kFromLastElem.Value)
 	kFromLastElemRec := findKFromLastRecr(l.Front(), 3, &WrapObj{0})
-	fmt.Println("Recursive Method : ",kFromLastElemRec.Value.(int))
+	fmt.Println("Recursive Method : ", kFromLastElemRec.Value.(int))
 }
+
 //Iterative function to find the kth from last element
 func findKFromLast(l *list.List, k int) *list.Element {
-    size := l.Len()
-    //Base condition. If the size of the list is less than k then kth element cannot be found
-    if size < k {
-        return nil
-    }
-    var elem *list.Element
-    elem = l.Front()
-    for i := 1; i < k; i++ {
-        elem = elem.Next()
-    }
-    var first *list.Element
-    for first = l.Front(); first != nil && elem != nil; elem, first = elem.Next(), first.Next() {
-    	//return the current node when current+k position is nil
-        if elem.Next() == nil {
-            return first
-        }
-    }
-    return nil
+	size := l.Len()
+	//Base condition. If the size of the list is less than k then kth element cannot be found
+	if size < k {
+		return nil
+	}
+	var elem *list.Element
+	elem = l.Front()
+	for i := 1; i < k; i++ {
+		elem = elem.Next()
+	}
+	var first *list.Element
+	for first = l.Front(); first != nil && elem != nil; elem, first = elem.Next(), first.Next() {
+		//return the current node when current+k position is nil
+		if elem.Next() == nil {
+			return first
+		}
+	}
+	return nil
 }
+
 //Object to store the count 
 type WrapObj struct {
 	count int
 }
+
 //recursive function to find the kth from last element
 func findKFromLastRecr(head *list.Element, k int, wrapper *WrapObj) *list.Element {
 	if head == nil {
